Phase 4b: extract NervesHub.Devices.Pinning sub-context
Second step of splitting the Devices god-module. Built on Phase 4a
(cleanup/phase-4a-devices-cacertificates).

Moves the five user-facing device-pinning functions (get_pinned_devices/1,
pin_device/2, unpin_device/2, device_pinned?/2, unpin_org_devices/2) out of
NervesHub.Devices into a new, documented NervesHub.Devices.Pinning module.

The PinnedDevice schema stays referenced by Devices for the non-pinning
pinned-status join and device cleanup, so only the dedicated pinning
operations moved. Function names unchanged; the 23 call sites across lib and
test were repointed, and the now-unused `Devices` alias was dropped from
Accounts and the orgs index LiveView.

Verification: mix format, credo (new module clean), dialyzer (0 unskipped),
full test suite (1396 passing).
